Feature/doc summarizer (#15)

+++ src/main/java/com/ayushsingh/doc_helper/features/product_features/dto/feature_product/BillingProductFeatureDetailsDto.java
@@ -15,7 +15,7 @@ public class BillingProductFeatureDetailsDto {
     private Long id;
     private Long productId;
     private Long featureId;
-    private Long enabledVersion;
+    private Integer enabledVersion;
     private Boolean enabled;
     private Integer priority;
     private Long quotaLimit;
